+++ packages/entities/entities-plugins/src/components/PluginConfigCard.vue
@@ -33,65 +33,66 @@
       @fetch:success="(entity: any) => $emit('fetch:success', entity)"
       @loading="(val: boolean) => $emit('loading', val)"
     >
-      <template #name="{ rowValue }">
+      <template #name="slotProps">
         <div class="name-cell-wrapper">
           <PluginIcon
             class="plugin-icon"
-            :name="rowValue"
+            :name="getPropValue('rowValue', slotProps)"
             :width="24"
           />
           <span class="info-name">
-            {{ pluginMetaData.getDisplayName(rowValue) }}
+            {{ pluginMetaData.getDisplayName(getPropValue('rowValue', slotProps)) }}
           </span>
         </div>
       </template>
 
-      <template #consumer="{ rowValue, row }">
-        <span v-if="!rowValue">–</span>
+      <template #consumer="slotProps">
+        <span v-if="!getPropValue('rowValue', slotProps)">–</span>
         <CopyUuid
           v-else
           data-testid="consumer-copy-uuid"
           :notify="() => {}"
           :success-tooltip="t('copy.success_tooltip')"
-          :tooltip="t('copy.tooltip', { label: row.label })"
+          :tooltip="t('copy.tooltip', { label: getPropValue('row', slotProps).label })"
           :truncated="false"
-          :uuid="rowValue.id"
+          :uuid="getPropValue('rowValue', slotProps).id"
         />
       </template>
-      <template #route="{ rowValue, row }">
-        <span v-if="!rowValue">–</span>
+
+      <template #route="slotProps">
+        <span v-if="!getPropValue('rowValue', slotProps)">–</span>
         <CopyUuid
           v-else
           data-testid="route-copy-uuid"
           :notify="() => {}"
           :success-tooltip="t('copy.success_tooltip')"
-          :tooltip="t('copy.tooltip', { label: row.label })"
+          :tooltip="t('copy.tooltip', { label: getPropValue('row', slotProps).label })"
           :truncated="false"
-          :uuid="rowValue.id"
+          :uuid="getPropValue('rowValue', slotProps).id"
         />
       </template>
-      <template #service="{ rowValue, row }">
-        <span v-if="!rowValue">–</span>
+      <template #service="slotProps">
+        <span v-if="!getPropValue('rowValue', slotProps)">–</span>
         <CopyUuid
           v-else
           data-testid="service-copy-uuid"
           :notify="() => {}"
           :success-tooltip="t('copy.success_tooltip')"
-          :tooltip="t('copy.tooltip', { label: row.label })"
+          :tooltip="t('copy.tooltip', { label: getPropValue('row', slotProps).label })"
           :truncated="false"
-          :uuid="rowValue.id"
+          :uuid="getPropValue('rowValue', slotProps).id"
         />
       </template>
-      <template #consumer_group="{ rowValue, row }">
-        <span v-if="!rowValue">–</span>
+      <template #consumer_group="slotProps">
+        <span v-if="!getPropValue('rowValue', slotProps)">–</span>
         <CopyUuid
           v-else
           data-testid="consumer-group-copy-uuid"
           :notify="() => { }"
           :success-tooltip="t('copy.success_tooltip')"
-          :tooltip="t('copy.tooltip', { label: row.label })"
+          :tooltip="t('copy.tooltip', { label: getPropValue('row', slotProps).label })"
           :truncated="false"
-          :uuid="rowValue.id"
+          :uuid="getPropValue('rowValue', slotProps).id"
         />
       </template>
     </EntityBaseConfigCard>
@@ -113,6 +114,7 @@ import {
   ConfigurationSchemaSection,
   useAxios,
   useErrors,
+  useHelpers,
 } from '@kong-ui-public/entities-shared'
 import { CopyUuid } from '@kong-ui-public/copy-uuid'
 import composables from '../composables'
@@ -169,6 +171,7 @@ const props = defineProps({
 const { i18n: { t } } = composables.useI18n()
 const pluginMetaData = composables.usePluginMetaData()
 const { setFieldType } = composables.usePluginHelpers()
+const { getPropValue } = useHelpers()
 
 const fetchUrl = computed<string>(
   () => endpoints.item[props.config.app]?.[props.scopedEntityType ? 'forEntity' : 'all']